+++ Python/NeonOcean.Debug/NeonOcean/Debug/Interactions/Support/Categories.py
@@ -1,2 +1,0 @@
-DebugID = 9828540712044566059  # type: int
-DebugModSettingsID = 7714482771736482519  # type: int