+++ Python/NeonOcean.S4.Debug/NeonOcean/S4/Debug/This.py
@@ -1,0 +1,8 @@
+from __future__ import annotations
+
+from NeonOcean.S4.Main import Mods
+
+try:
+	Mod = Mods.GetMod("NeonOcean.S4.Debug")  # type: Mods.Mod
+except Exception as e:
+	raise Exception("Cannot find self in mod list.") from e